requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
drf-yasg==1.20.0
gevent==20.12.1
greenlet==0.4.17
grequests==0.6.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
zope.event==4.5.0
zope.interface==5.2.0
